+++ CHANGELOG.md
@@ -8,13 +8,15 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 ## [Unreleased](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/compare/1.1.1...main)
 
 ### Fixed
-- rename results' index from 'raster_id_100m' to 'index' to support the grid value visualization in dashboard([41](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/41))
+- rename results' index from 'raster_id_100m' to 'index' to support the grid value visualization in dashboard ([41](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/41))
 
 ### Changed
 - Introduced artifact tags ([40](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/40))
+- Changed the database connection to read tables from the `census_de` schema ([#32](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/32))
 
 ### Added
 - Maps of dominant building ages and energy carriers ([28](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/28))
+- Added tests for gridded artifact functions, several functions in utils, and get_clipped_census_grid ([24](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/24), [38](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/issues/38))
 
 ## [1.1.1](https://gitlab.heigit.org/climate-action/plugins/heating-emissions/-/releases/1.1.1) - 2025-06-04